requirements.txt
asgiref==3.4.1
astroid==2.8.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
django-4-jet==1.0.9
django-ckeditor==6.1.0
django-cors-headers
django-crispy-forms==1.13.0
django-environ==0.8.0
django-filter==23.2
django-js-asset==1.2.2
djangorestframework==3.14.0
idna==3.4
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.6.0
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.6.1
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==2.3.0
psycopg2==2.9.6
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.12.1